+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 ## Changelog
 
-##### not released
+##### hugo-search 1.1.0 - (04 December 2016)
 
     * Code clean up
     * Unregister index before closing
+    * Fix build due API changes in Hugo (page.RelPermalink())
+    * Do not index special pages such as taxonomies and search results
+    * Revert Makefile to work with make.exe from http://win-builds.org
 
 ##### hugo-search 1.0.2 - (08 September 2016)